requirements.txt
PyJWT>=2.8.0
flask-cors>=4.0.0
google-auth>=4.0.0
